修复 Bilibili WebSocket 栈溢出问题
- 修正头部协议版本（保持为1）和JSON中protover（保持为2）的混淆
- 之前错误地将头部协议版本改为3导致栈溢出
- 简化DanmakuMessage解析逻辑，移除递归处理
- 暂时跳过Brotli压缩数据的处理

+++ ChatCore/Models/BiliBili/BiliBiliChatUser.cs
@@ -162,7 +162,7 @@ namespace ChatCore.Models.Bilibili
 			this.IsModerator = IsModerator == 1;
 		}
 
-		public void SetIsFan(int MedalOwnerRoomId, int BoardcasterUid = 0) {
+		public void SetIsFan(int MedalOwnerRoomId, long BoardcasterUid = 0) {
 			IsFan = (MedalOwnerRoomId == -1 && BoardcasterUid == -1) || (MedalOwnerRoomId == BilibiliService._roomID) || (BoardcasterUid == BilibiliService._userID);
 		}
 
@@ -175,7 +175,7 @@ namespace ChatCore.Models.Bilibili
 			HonorLevel = Level;
 		}
 
-		public void SetMedal(int Level = 0, string MedalName = "", int GuardLevel = 0, int MedalOwnerRoomId = 0, int BoardcasterUid = 0)
+		public void SetMedal(int Level = 0, string MedalName = "", int GuardLevel = 0, int MedalOwnerRoomId = 0, long BoardcasterUid = 0)
 		{
 			var badgeList = new List<IChatBadge>();
 			var newBadge = new BilibiliChatBadge();
@@ -194,7 +194,7 @@ namespace ChatCore.Models.Bilibili
 			SetIsFan(MedalOwnerRoomId, BoardcasterUid);
 		}
 
-		public void SetMedal(int Level, string MedalName, bool ColorInInt, string[] ColorData, int GuardLevel, int MedalOwnerRoomId, int BoardcasterUid) {
+		public void SetMedal(int Level, string MedalName, bool ColorInInt, string[] ColorData, int GuardLevel, int MedalOwnerRoomId, long BoardcasterUid) {
 			var badgeList = new List<IChatBadge>();
 			var newBadge = new BilibiliChatBadge();
 			newBadge.Name = MedalName;
@@ -312,7 +312,7 @@ namespace ChatCore.Models.Bilibili
 				Avatar = string.Empty;
 				return;
 			}
-			
+
 			var avatarImageUrl = await GetUserInfoAsync(Id) + @"@128w_128h.png";
 			var scale = 3;
 			var sb = new StringBuilder(SVG_FRAME);
